+++ packages/common-ethereum/src/project/load.ts
@@ -1,33 +1,10 @@
 // Copyright 2020-2023 SubQuery Pte Ltd authors & contributors
 // SPDX-License-Identifier: GPL-3.0
 
-import fs from 'fs';
-import path from 'path';
-import {loadFromJsonOrYaml} from '@subql/common';
 import {EthereumProjectManifestVersioned, VersionedProjectManifest} from './versioned';
 
 export function parseEthereumProjectManifest(raw: unknown): EthereumProjectManifestVersioned {
   const projectManifest = new EthereumProjectManifestVersioned(raw as VersionedProjectManifest);
-  projectManifest.validate();
-  return projectManifest;
-}
-
-export function loadEthereumProjectManifest(file: string): EthereumProjectManifestVersioned {
-  let manifestPath = file;
-  if (fs.existsSync(file) && fs.lstatSync(file).isDirectory()) {
-    const yamlFilePath = path.join(file, 'project.yaml');
-    const jsonFilePath = path.join(file, 'project.json');
-    if (fs.existsSync(yamlFilePath)) {
-      manifestPath = yamlFilePath;
-    } else if (fs.existsSync(jsonFilePath)) {
-      manifestPath = jsonFilePath;
-    } else {
-      throw new Error(`Could not find project manifest under dir ${file}`);
-    }
-  }
-
-  const doc = loadFromJsonOrYaml(manifestPath);
-  const projectManifest = new EthereumProjectManifestVersioned(doc as VersionedProjectManifest);
   projectManifest.validate();
   return projectManifest;
 }
