+++ packages/common-ethereum/src/project/versioned/ProjectManifestVersioned.ts
@@ -42,7 +42,7 @@ export class EthereumProjectManifestVersioned implements IEthereumProjectManifes
   }
 
   toDeployment(): string | undefined {
-    return this._impl.toDeployment();
+    return this._impl.deployment.toYaml();
   }
 
   validate(): void {
