+++ packages/node/src/configure/SubqueryProject.ts
@@ -7,7 +7,6 @@ import { validateSemver } from '@subql/common';
 import {
   parseEthereumProjectManifest,
   SubqlEthereumDataSource,
-  EthereumBlockFilter,
   ProjectManifestV1_0_0Impl,
   isRuntimeDs,
   EthereumHandlerKind,
@@ -26,7 +25,6 @@ import {
   CustomDatasourceTemplate,
 } from '@subql/types-ethereum';
 import { buildSchemaFromString } from '@subql/utils';
-import Cron from 'cron-converter';
 import { GraphQLSchema } from 'graphql';
 import { updateDatasourcesFlare } from '../utils/project';
 
@@ -37,13 +35,6 @@ export type EthereumProjectDs = SubqlProjectDs<SubqlEthereumDataSource>;
 export type EthereumProjectDsTemplate =
   | SubqlProjectDs<RuntimeDatasourceTemplate>
   | SubqlProjectDs<CustomDatasourceTemplate>;
-
-export type SubqlProjectBlockFilter = EthereumBlockFilter & {
-  cronSchedule?: {
-    schedule: Cron.Seeker;
-    next: number;
-  };
-};
 
 const NOT_SUPPORT = (name: string) => {
   throw new Error(`Manifest specVersion ${name} is not supported`);
